fix(table): styles and layout for expandable rows
