+++ elements/pf-table/demo/expandable-rows.html
@@ -26,7 +26,7 @@
 
     <pf-td slot="expansion">
 
-      <p>Details of the dependency: <strong>io.quarkus:quarkus-hibernate-orm</strong></pf-td>
+      <p>Details of the dependency: <strong>io.quarkus:quarkus-hibernate-orm</strong></p>
       <pf-table>
         <pf-tr>
           <pf-th>Severity</pf-th>
@@ -83,7 +83,6 @@
     </pf-td>
 
   </pf-tr>
-  <pf-tr id="ex-expandable-row" expandable></pf-tr>
 
   <pf-tr id="2">
     <pf-td>
