+++ elements/pf-table/pf-tbody.css
@@ -1,40 +1,3 @@
-:host([expandable]),
-:host([expandable][expanded]) ::slotted(*) {
-  position: relative;
-}
-
-:host([expandable][expanded]) ::slotted(*)::before {
-  position: absolute;
-  inset-block-start: -1px;
-  inset-inline-end: 0;
-  inset-block-end: -1px;
-  inset-inline-start: 0;
-  content: "";
-  border-inline-start: 3px solid #06c;
-  pointer-events: none;
-}
-
-:host([expandable]) ::slotted(pf-table) {
-  margin-block-end: -1px;
-}
-
-#toggle-cell {
-  z-index: 2;
-  position: absolute;
-  inset-inline-start: 0;
-  inset-block-start: 1.5rem;
-  padding: 0;
-}
-
-#toggle-icon {
-  transition: .2s ease-in 0s;
-  vertical-align: -0.125em;
-}
-
-:host([expanded]) #toggle-icon {
-  transform: rotate(90deg);
-}
-
 @media (min-width: 768px) {
   :host {
     display: grid;
