+++ elements/pf-table/pf-tr.css
@@ -20,8 +20,47 @@
   display: contents;
 }
 
-:host([expandable]) {
+:host([expandable]),
+:host([expandable][expanded]) #expansion::slotted(*) {
+  position: relative;
+}
+
+:host([expandable][expanded]) #expansion::slotted(*)::before {
+  position: absolute;
+  inset-block-start: -1px;
+  inset-inline-end: 0;
+  inset-block-end: -1px;
+  inset-inline-start: 0;
+  content: "";
+  border-inline-start: 3px solid #06c;
+  pointer-events: none;
+}
+
+:host([expandable]) #expansion::slotted(pf-table) {
+  margin-block-end: -1px;
+}
+
+#toggle-cell {
+  z-index: 2;
+  position: absolute;
+  inset-inline-start: 0;
+  inset-block-start: 1.5rem;
+  padding: 0;
+}
+
+#toggle-icon {
+  transition: .2s ease-in 0s;
+  vertical-align: -0.125em;
+}
+
+:host([expanded]) #toggle-icon {
+  transform: rotate(90deg);
+}
+
+:host([expandable]) #expansion {
   grid-auto-flow: row;
   grid-template-columns: auto;
-  padding: 0 1.5rem;
+  padding: 0px 1.5rem;
+  display: grid;
+  grid-column: 1/-1;
 }
